refactor: split grammar models and functions
The single 'grammar.py' model file was getting big enough with just the
models for the parts and the project itself. Split it up so we can grow
these in a more "sane" fashion.

The same applies to the tests: 'test_application.py' is already huge so
move the tests that check grammar processing to a new
'test_process_grammar.py'.
