+++ craft_application/models/grammar/__init__.py
@@ -1,0 +1,22 @@
+# This file is part of craft-application.
+#
+# Copyright 2023-2024 Canonical Ltd.
+#
+# This program is free software: you can redistribute it and/or modify it
+# under the terms of the GNU Lesser General Public License version 3, as
+# published by the Free Software Foundation.
+#
+# This program is distributed in the hope that it will be useful, but WITHOUT
+# ANY WARRANTY; without even the implied warranties of MERCHANTABILITY,
+# SATISFACTORY QUALITY, or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+# General Public License for more details.
+#
+# You should have received a copy of the GNU Lesser General Public License along
+# with this program.  If not, see <http://www.gnu.org/licenses/>.
+"""Models for grammar validation."""
+
+from ._parts import get_grammar_aware_part_keywords
+from ._project import GrammarAwareProject
+
+
+__all__ = ["GrammarAwareProject", "get_grammar_aware_part_keywords"]
