+++ craft_application/models/grammar/_base.py
@@ -1,0 +1,30 @@
+# This file is part of craft-application.
+#
+# Copyright 2023-2024 Canonical Ltd.
+#
+# This program is free software: you can redistribute it and/or modify it
+# under the terms of the GNU Lesser General Public License version 3, as
+# published by the Free Software Foundation.
+#
+# This program is distributed in the hope that it will be useful, but WITHOUT
+# ANY WARRANTY; without even the implied warranties of MERCHANTABILITY,
+# SATISFACTORY QUALITY, or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+# General Public License for more details.
+#
+# You should have received a copy of the GNU Lesser General Public License along
+# with this program.  If not, see <http://www.gnu.org/licenses/>.
+"""Bases for grammar-related models."""
+
+import pydantic
+
+from craft_application.models.base import alias_generator
+
+
+class _GrammarAwareModel(pydantic.BaseModel):
+    class Config:
+        """Default configuration for grammar-aware models."""
+
+        validate_assignment = True
+        extra = pydantic.Extra.allow  # verify only grammar-aware parts
+        alias_generator = alias_generator
+        allow_population_by_field_name = True
